requirements.txt
qdrant-client==1.7.3
langchain==0.1.5
langchain-community==0.0.20
langchain-openai==0.0.6
openai==1.12.0
streamlit==1.31.0
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.2.0
sentence-transformers==2.2.2
python-dotenv==1.0.1
chromadb==0.4.22
tiktoken==0.5.2